move in separate type
