+++ rpc/api_contract_test.go
@@ -5,49 +5,6 @@ import (
 	"testing"
 )
 
-// TestCodeAt tests code for a contract instance. This will be deprecated.
-func TestCodeAt(t *testing.T) {
-	testConfig := beforeEach(t)
-
-	type testSetType struct {
-		ContractHash      string
-		ExpectedBytecode0 string
-	}
-	testSet := map[string][]testSetType{
-		"mock": {
-			{
-				ContractHash:      "0xdeadbeef",
-				ExpectedBytecode0: "0xdeadbeef",
-			},
-		},
-		"testnet": {
-			{
-				ContractHash:      "0x6fbd460228d843b7fbef670ff15607bf72e19fa94de21e29811ada167b4ca39",
-				ExpectedBytecode0: "0x480680017fff8000",
-			},
-		},
-		"mainnet": {
-			{
-				ContractHash:      "0x3b4be7def2fc08589348966255e101824928659ebb724855223ff3a8c831efa",
-				ExpectedBytecode0: "0x40780017fff7fff",
-			},
-		},
-	}[testEnv]
-
-	for _, test := range testSet {
-		code, err := testConfig.client.CodeAt(context.Background(), test.ContractHash)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if code == nil || len(code.Bytecode) == 0 {
-			t.Fatal("code should exist")
-		}
-		if code.Bytecode[0] != test.ExpectedBytecode0 {
-			t.Fatalf("expecting bytecode[0] %s, got %s", test.ExpectedBytecode0, code.Bytecode[0])
-		}
-	}
-}
-
 // TestClassAt tests code for a class.
 func TestClassAt(t *testing.T) {
 	testConfig := beforeEach(t)
@@ -82,11 +39,8 @@ func TestClassAt(t *testing.T) {
 		if err != nil {
 			t.Fatal(err)
 		}
-		if class == nil || class.Program == nil {
+		if class == nil || class.Program == "" {
 			t.Fatal("code should exist")
-		}
-		if _, ok := class.Program.(string); !ok {
-			t.Fatalf("program should cast to a 'string', instead %T", class.Program)
 		}
 	}
 }
@@ -140,24 +94,28 @@ func TestClass(t *testing.T) {
 	testConfig := beforeEach(t)
 
 	type testSetType struct {
+		BlockIDOption     BlockIDOption
 		ClassHash         string
 		ExpectedOperation string
 	}
 	testSet := map[string][]testSetType{
 		"mock": {
 			{
+				BlockIDOption:     WithBlockIDTag("pending"),
 				ClassHash:         "0xdeadbeef",
 				ExpectedOperation: "0xdeadbeef",
 			},
 		},
 		"testnet": {
 			{
+				BlockIDOption:     WithBlockIDTag("pending"),
 				ClassHash:         "0x493af3546940eb96471cf95ae3a5aa1286217b07edd1e12d00143010ca904b1",
 				ExpectedOperation: "0x40780017fff7fff",
 			},
 		},
 		"mainnet": {
 			{
+				BlockIDOption:     WithBlockIDTag("pending"),
 				ClassHash:         "0x4c53698c9a42341e4123632e87b752d6ae470ddedeb8b0063eaa2deea387eeb",
 				ExpectedOperation: "0x40780017fff7fff",
 			},
@@ -165,15 +123,12 @@ func TestClass(t *testing.T) {
 	}[testEnv]
 
 	for _, test := range testSet {
-		class, err := testConfig.client.Class(context.Background(), test.ClassHash)
+		class, err := testConfig.client.Class(context.Background(), test.BlockIDOption, test.ClassHash)
 		if err != nil {
 			t.Fatal(err)
 		}
-		if class == nil || class.Program == nil {
+		if class == nil || class.Program == "" {
 			t.Fatal("code should exist")
-		}
-		if _, ok := class.Program.(string); !ok {
-			t.Fatalf("program should cast to a 'string', instead %T", class.Program)
 		}
 	}
 }
